requirements.txt
python-telegram-bot
qdrant-client
google-generativeai
python-dotenv
sentence-transformers
